Decouple CheckForConflicts from DuckLake Transaction
