+++ src/include/storage/ducklake_transaction_state.hpp
@@ -21,7 +21,14 @@ public:
 	~DuckLakeTransactionState();
 
 	void Commit(DuckLakeSnapshot transaction_snapshot, const TransactionChangeInformation &transaction_changes,
-	            const DuckLakeRetryConfig &retry_config);
+	            const DuckLakeRetryConfig &retry_config,
+	            const std::function<unique_ptr<QueryResult>(string)> &conflict_query_executor);
+
+	SnapshotAndStats CheckForConflicts(DuckLakeSnapshot transaction_snapshot,
+	                                   const TransactionChangeInformation &changes,
+	                                   const std::function<unique_ptr<QueryResult>(string)> &executor);
+	void CheckForConflicts(const TransactionChangeInformation &changes, const SnapshotChangeInformation &other_changes,
+	                       DuckLakeSnapshot transaction_snapshot) const;
 
 public:
 	DuckLakeTransaction &transaction;
